Generate daDemo destructor thunks natively (#1926)
* Use native instrument bank ownership

* Record instrument bank ownership provenance

* Claim daDemo destructor ownership

* Generate daDemo destructor thunks natively

* Record native destructor thunk evidence

* Close daDemo destructor ownership claim

* Keep generated and coordination ledgers out of lifecycle slice

* Keep generated and coordination ledgers out of lifecycle slice
