+++ config/match_provenance.jsonl
@@ -535,3 +535,5 @@
 {"id":"arm9:0x0204e40c","module":"arm9","addr":33874956,"name":"_ZN22ExpandingHeapAllocator8FreeNodeEP10MemoryNodePNS0_6TargetE","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN22ExpandingHeapAllocator8FreeNodeEP10MemoryNodePNS0_6TargetE.cpp"}
 {"id":"arm9:0x02050c78","module":"arm9","addr":33885304,"name":"_ZN5Sound17InfoSequenceEntry9GetWithIDEj","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN5Sound17InfoSequenceEntry9GetWithIDEj.cpp"}
 {"id":"arm9:0x02050bb0","module":"arm9","addr":33885104,"name":"_ZN5Sound23InfoInstrumentBankEntry9GetWithIDEj","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN5Sound23InfoInstrumentBankEntry9GetWithIDEj.cpp"}
+{"id":"ov002:0x020f8838","module":"ov002","addr":34572344,"name":"_ZThn80_N8daDemo_c10anmModel_cD0Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZThn80_N8daDemo_c10anmModel_cD0Ev.cpp"}
+{"id":"ov002:0x020f8848","module":"ov002","addr":34572360,"name":"_ZThn80_N8daDemo_c10anmModel_cD1Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZThn80_N8daDemo_c10anmModel_cD1Ev.cpp"}
